バグ修正と微調整と仮UI

+++ Button.cpp
@@ -8,20 +8,20 @@ void Button::mainProcess(bool isSelected, bool shadow ,int yoffset)
     {
         ActualPosition = VGet(BoxSizeChange(ActualPosition.x, SelectedPosition1.x, false, NotSelectedPosition1.x), BoxSizeChange(ActualPosition.y, SelectedPosition1.y, false, NotSelectedPosition1.y),1);
         ActualPosition2 = VGet(BoxSizeChange(ActualPosition2.x, SelectedPosition2.x, true, NotSelectedPosition2.x), BoxSizeChange(ActualPosition2.y, SelectedPosition2.y, true, NotSelectedPosition2.y), 1);
-        DrawBox(Button::ActualPosition.x, Button::ActualPosition.y, Button::ActualPosition2.x, Button::ActualPosition2.y, Button::SelectedColor, 1);
-        DrawBox(Button::ActualPosition.x, Button::ActualPosition.y, Button::ActualPosition2.x, Button::ActualPosition2.y, GetColor(255,255,255), 0,3);
-        textPositionSet(ActualPosition.x, ActualPosition2.x, buttonText, BiggerFontHandle, SORT_CENTER, ActualPosition.y + yoffset*0.6, true, GetColor(255, 255, 255));
+        DrawBox((int)Button::ActualPosition.x, (int)Button::ActualPosition.y, (int)Button::ActualPosition2.x, (int)Button::ActualPosition2.y, Button::SelectedColor, 1);
+        DrawBox((int)Button::ActualPosition.x, (int)Button::ActualPosition.y, (int)Button::ActualPosition2.x, (int)Button::ActualPosition2.y, GetColor(255,255,255), 0,3);
+        textPositionSet((int)ActualPosition.x, (int)ActualPosition2.x, buttonText, BiggerFontHandle, SORT_CENTER, (int)ActualPosition.y + (int)yoffset*0.6f, true, GetColor(255, 255, 255));
     }
     else
     {
         ActualPosition = VGet(BoxSizeChange(ActualPosition.x, NotSelectedPosition1.x, true, SelectedPosition1.x), BoxSizeChange(ActualPosition.y, NotSelectedPosition1.y, true, SelectedPosition1.y), 1);
         ActualPosition2 = VGet(BoxSizeChange(ActualPosition2.x, NotSelectedPosition2.x, false, SelectedPosition2.x), BoxSizeChange(ActualPosition2.y, NotSelectedPosition2.y, false, SelectedPosition2.y), 1);
-        DrawBox(Button::ActualPosition.x, Button::ActualPosition.y, Button::ActualPosition2.x, Button::ActualPosition2.y, Button::NotSelectedColor, 1);
-        textPositionSet(ActualPosition.x, ActualPosition2.x, buttonText, fontHandle, SORT_CENTER, ActualPosition.y + yoffset*0.6, true, GetColor(122,122,122));
+        DrawBox((int)Button::ActualPosition.x, (int)Button::ActualPosition.y, (int)Button::ActualPosition2.x, (int)Button::ActualPosition2.y, Button::NotSelectedColor, 1);
+        textPositionSet((int)ActualPosition.x, (int)ActualPosition2.x, buttonText, fontHandle, SORT_CENTER, (int)ActualPosition.y + (int)yoffset*0.6f, true, GetColor(122,122,122));
     }
     
 }
-void Button::SetButtonPosition(VECTOR CenterPosition, int width, int height, float reductionRatio)
+void Button::SetButtonPosition(VECTOR CenterPosition, float width, float height, float reductionRatio)
 {
     width /= 2;
     height /= 2;
